Remove "she/her/Veronica/Josie" from Jeep posts
The Jeep is a vehicle, not a person. Stop engendering it. Additionally,
in an effort to improve my own communication and remove things that are
subtly sexist (I think people refer to them as micro-aggressions?), I'm
stripping them out.

I want this to be inviting. I need to improve how I speak and write.